Fix detach surface cts test run timeout issue
The issue is detach surface test run timeout, cause is in it keeps
reallocating surface when the allocated surface is not in cached
surfaces pool created during decoder initialization stage.

Solution is 1) Remove code to check if allocated surfce is in
cache. 2) Change to use IGBA instead of BUFFERQUEUE as output
surface allocator for android 15.

Tracked-On: OAM-129505
Signed-off-by: Lina Sun <lina.sun@intel.com>